fix compilation for 32 bit architectures

+++ src/fb/FbException.h
@@ -15,7 +15,9 @@
 #ifndef DBWRAP_FB_FBEXCEPTION_H_
 #define DBWRAP_FB_FBEXCEPTION_H_
 
+#include <cstdint>
 #include <stdexcept>
+
 
 namespace fb {
 
@@ -23,7 +25,7 @@ class FbException : public std::runtime_error
 {
 public:
     /** status should be an ISC_STATUS_ARRAY from ibase.h */
-    FbException(const char *operation, const long *status);
+    FbException(const char *operation, const intptr_t *status);
     FbException(const FbException&) = default;
 
     virtual ~FbException() noexcept;
